remove code paths related to "num_local_schedulers"; convert "address_info" into a property
create "runner" module

create "session" module
