+++ python/ray/__init__.py
@@ -57,7 +57,7 @@ from ray.worker import (error_info, init, connect, disconnect, get, put, wait,
                         is_initialized)  # noqa: E402
 from ray.worker import (SCRIPT_MODE, WORKER_MODE, LOCAL_MODE,
                         PYTHON_MODE)  # noqa: E402
-from ray.worker import global_state  # noqa: E402
+from ray.worker import get_global_state  # noqa: E402
 import ray.internal  # noqa: E402
 # We import ray.actor because some code is run in actor.py which initializes
 # some functions in the worker.
@@ -72,7 +72,7 @@ __all__ = [
     "remote", "profile", "actor", "method", "get_gpu_ids", "get_resource_ids",
     "get_webui_url", "register_custom_serializer", "shutdown",
     "is_initialized", "SCRIPT_MODE", "WORKER_MODE", "LOCAL_MODE",
-    "PYTHON_MODE", "global_state", "ObjectID", "_config", "__version__",
+    "PYTHON_MODE", "get_global_state", "ObjectID", "_config", "__version__",
     "internal"
 ]
