+++ python/ray/parameter.py
@@ -2,7 +2,9 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
+import json
 import logging
+import os
 
 import ray.ray_constants as ray_constants
 
@@ -11,11 +13,6 @@ class RayParams(object):
     """A class used to store the parameters used by Ray.
 
     Attributes:
-        address_info (dict): A dictionary with address information for
-            processes in a partially-started Ray cluster. If
-            start_ray_local=True, any processes not in this dictionary will be
-            started. If provided, an updated address_info dictionary will be
-            returned to include processes that are newly started.
         start_ray_local (bool): If True then this will start any processes not
             already in address_info, including Redis, a global scheduler, local
             scheduler(s), object store(s), and worker(s). It will also kill
@@ -32,12 +29,13 @@ class RayParams(object):
             ignored.
         redis_shard_ports: A list of the ports to use for the non-primary Redis
             shards.
+        use_credis: If True, additionally load the chain-replicated libraries
+            into the redis servers.  Defaults to None, which means its value is
+            set by the presence of "RAY_USE_NEW_GCS" in os.environ.
         num_cpus (int): Number of cpus the user wishes all local schedulers to
             be configured with.
         num_gpus (int): Number of gpus the user wishes all local schedulers to
             be configured with.
-        num_local_schedulers (int): The number of local schedulers to start.
-            This is only provided if start_ray_local is True.
         resources: A dictionary mapping the name of a resource to the quantity
             of that resource available.
         object_store_memory: The amount of memory (in bytes) to start the
@@ -46,12 +44,12 @@ class RayParams(object):
             to use, or None for no limit. Once the limit is exceeded, redis
             will start LRU eviction of entries. This only applies to the
             sharded redis tables (task and object tables).
-        object_manager_ports (list): A list of the ports to use for the object
-            managers. There should be one per object manager being started on
-            this node (typically just one).
-        node_manager_ports (list): A list of the ports to use for the node
-            managers. There should be one per node manager being started on
-            this node (typically just one).
+        object_manager_port (int): The port to use for the object managers.
+            There should be one per object manager being started on this node
+            (typically just one).
+        node_manager_port (int): The port to use for the node managers.
+            There should be one per node manager being started on this node
+            (typically just one).
         node_ip_address (str): The IP address of the node that we are on.
         object_id_seed (int): Used to seed the deterministic generation of
             object IDs. The same value can be used across multiple runs of the
@@ -92,19 +90,18 @@ class RayParams(object):
     """
 
     def __init__(self,
-                 address_info=None,
                  start_ray_local=False,
                  redis_address=None,
                  num_cpus=None,
                  num_gpus=None,
-                 num_local_schedulers=None,
                  resources=None,
                  object_store_memory=None,
                  redis_max_memory=None,
                  redis_port=None,
                  redis_shard_ports=None,
-                 object_manager_ports=None,
-                 node_manager_ports=None,
+                 use_credis=None,
+                 object_manager_port=None,
+                 node_manager_port=None,
                  node_ip_address=None,
                  object_id_seed=None,
                  num_workers=None,
@@ -127,35 +124,57 @@ class RayParams(object):
                  include_log_monitor=None,
                  autoscaling_config=None,
                  _internal_config=None):
-        self.address_info = address_info
+        if num_workers is not None:
+            raise Exception(
+                "The 'num_workers' argument is deprecated. Please use "
+                "'num_cpus' instead.")
         self.start_ray_local = start_ray_local
         self.object_id_seed = object_id_seed
-        self.redis_address = redis_address
+
+        # Node IP address
+        self.node_ip_address = node_ip_address
+
+        # Resources settings.
         self.num_cpus = num_cpus
         self.num_gpus = num_gpus
-        self.num_local_schedulers = num_local_schedulers
         self.resources = resources
-        self.object_store_memory = object_store_memory
-        self.redis_max_memory = redis_max_memory
+
+        # Redis settings
+        self.redis_address = redis_address
         self.redis_port = redis_port
+        self.redis_password = redis_password
+        self.redis_max_clients = redis_max_clients
+        self.redis_max_memory = redis_max_memory
+        self.num_redis_shards = num_redis_shards
         self.redis_shard_ports = redis_shard_ports
-        self.object_manager_ports = object_manager_ports
-        self.node_manager_ports = node_manager_ports
-        self.node_ip_address = node_ip_address
+        if use_credis is None:
+            self.use_credis = ("RAY_USE_NEW_GCS" in os.environ)
+        else:
+            self.use_credis = use_credis
+
+        # Plasma object store settings.
+        self.plasma_directory = plasma_directory
+        self.plasma_store_socket_name = plasma_store_socket_name
+        self.object_store_memory = object_store_memory
+        self.huge_pages = huge_pages
+
+        # Raylet settings
+        self.raylet_socket_name = raylet_socket_name
+        self.object_manager_port = object_manager_port
+        self.node_manager_port = node_manager_port
         self.num_workers = num_workers
+
+        # WebUI settings
+        self.include_webui = include_webui
+
+        # Logging settings
+        self.redirect_output = redirect_output
+        self.redirect_worker_output = redirect_worker_output
+
+        self.worker_path = worker_path
         self.local_mode = local_mode
         self.driver_mode = driver_mode
-        self.redirect_worker_output = redirect_worker_output
-        self.redirect_output = redirect_output
-        self.num_redis_shards = num_redis_shards
-        self.redis_max_clients = redis_max_clients
-        self.redis_password = redis_password
-        self.plasma_directory = plasma_directory
-        self.worker_path = worker_path
-        self.huge_pages = huge_pages
-        self.include_webui = include_webui
-        self.plasma_store_socket_name = plasma_store_socket_name
-        self.raylet_socket_name = raylet_socket_name
+
         self.temp_dir = temp_dir
         self.include_log_monitor = include_log_monitor
         self.autoscaling_config = autoscaling_config
@@ -173,6 +192,11 @@ class RayParams(object):
             else:
                 raise ValueError("Invalid RayParams parameter in"
                                  " update: %s" % arg)
+
+    @property
+    def config(self):
+        return json.loads(
+            self._internal_config) if self._internal_config else None
 
     def update_if_absent(self, **kwargs):
         """Update the settings when the target fields are None.
