--- python/ray/plasma/__init__.py
@@ -1,7 +1,0 @@
-from __future__ import absolute_import
-from __future__ import division
-from __future__ import print_function
-
-from ray.plasma.plasma import start_plasma_store, DEFAULT_PLASMA_STORE_MEMORY
-
-__all__ = ["start_plasma_store", "DEFAULT_PLASMA_STORE_MEMORY"]
